# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "film-production-planner",
  "version": "1.0.0"
 },
 "paths": {
...
  "/v1/scripts/parse": {
   "post": {
    "tags": [
     "breakdown"
    ],
    "summary": "Read a screenplay into scenes, locations and cast. The result is a draft breakdown that a person must approve before anything is scheduled",
    "operationId": "parse_v1_scripts_parse_post",
    "parameters": [
     {
      "name": "Idempotency-Key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ParseIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/productions/{production_id}/breakdown:approve": {
   "post": {
    "tags": [
     "breakdown"
    ],
    "summary": "(+) A person confirms the parsed breakdown. Scheduling is refused until this is done",
    "operationId": "approve_v1_productions__production_id__breakdown_approve_post",
    "parameters": [
     {
      "name": "production_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Production Id"
      }
     },
     {
      "name": "Idempotency-Key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/availability": {
   "post": {
    "tags": [
     "cast"
    ],
    "summary": "Set the dates a cast member cannot work (replaces the earlier list). New scenarios start from these",
    "operationId": "availability_v1_availability_post",
    "parameters": [
     {
      "name": "Idempotency-Key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AvailabilityIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/scenarios": {
   "post": {
    "tags": [
     "scenarios"
    ],
    "summary": "Create a what-if: its own cast availability and weather. Solving it does not touch any other scenario",
    "operationId": "create_scenario_v1_scenarios_post",
    "parameters": [
     {
      "name": "Idempotency-Key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ScenarioIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/schedules/optimize": {
   "post": {
    "tags": [
     "schedule"
    ],
    "summary": "Solve a scenario: every scene on a shoot day at the lowest cost under the rules, staying close to the scenario it is based on (job)",
    "operationId": "optimize_v1_schedules_optimize_post",
    "parameters": [
     {
      "name": "Idempotency-Key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OptimizeIn"
       }
      }
     }
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/schedules/{scenario_id}": {
   "get": {
    "tags": [
     "schedule"
    ],
    "summary": "(+) The stripboard: each shoot day with its scenes in order",
    "operationId": "stripboard_v1_schedules__scenario_id__get",
    "parameters": [
     {
      "name": "scenario_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Scenario Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/budget/variance": {
   "get": {
    "tags": [
     "budget"
    ],
    "summary": "A scenario's cost by category against the budget, and against another scenario",
    "operationId": "variance_v1_budget_variance_get",
    "parameters": [
     {
      "name": "scenario_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Scenario Id"
      }
     },
     {
      "name": "against",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Against"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/call-sheets/{date}": {
   "get": {
    "tags": [
     "schedule"
    ],
    "summary": "The call sheet for one shoot date in a scenario: scenes in order, cast calls, daylight and rain risk",
    "operationId": "call_sheet_v1_call_sheets__date__get",
    "parameters": [
     {
      "name": "date",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "date",
       "title": "Date"
      }
     },
     {
      "name": "scenario_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Scenario Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
   "AvailabilityIn": {
    "properties": {
     "production_id": {
      "type": "string",
      "format": "uuid",
      "title": "Production Id"
     },
     "cast": {
      "type": "string",
      "maxLength": 80,
      "minLength": 1,
      "title": "Cast"
     },
     "unavailable": {
      "items": {
       "type": "string",
       "format": "date"
      },
      "type": "array",
      "maxItems": 200,
      "title": "Unavailable"
     }
    },
    "type": "object",
    "required": [
     "production_id",
     "cast",
     "unavailable"
    ],
    "title": "AvailabilityIn"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "OptimizeIn": {
    "properties": {
     "scenario_id": {
      "type": "string",
      "format": "uuid",
      "title": "Scenario Id"
     },
     "seconds": {
      "type": "number",
      "maximum": 60.0,
      "minimum": 1.0,
      "title": "Seconds",
      "description": "Time the solver may spend; it returns the best schedule found",
      "default": 12
     }
    },
    "type": "object",
    "required": [
     "scenario_id"
    ],
    "title": "OptimizeIn"
   },
   "ParseIn": {
    "properties": {
     "title": {
      "type": "string",
      "maxLength": 120,
      "minLength": 1,
      "title": "Title"
     },
     "source": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^demo-screenplay$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source",
      "description": "Use the built-in synthetic screenplay instead of sending text"
     },
     "text": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 600000
       },
       {
        "type": "null"
       }
      ],
      "title": "Text",
      "description": "A screenplay in standard format"
     },
     "start_date": {
      "type": "string",
      "format": "date",
      "title": "Start Date"
     },
     "shoot_days": {
      "type": "integer",
      "maximum": 80.0,
      "minimum": 1.0,
      "title": "Shoot Days",
      "description": "Weekdays available for shooting, from the start date",
      "default": 24
     },
     "budget": {
      "type": "number",
      "maximum": 1000000000.0,
      "minimum": 0.0,
      "title": "Budget"
     }
    },
    "type": "object",
    "required": [
     "title",
     "start_date",
     "budget"
    ],
    "title": "ParseIn"
   },
...
  }
 }
}